Fixed <rdar://problem/4008338> REGRESSION (125-178): opacity style not working, breaks fading images on okcupid.com
	There was a long standing bug in cssText().  Double value were always cast int!  So,
	opacity values values were incorrectly converted to text, i.e. 0.75 became 0.

	The problem was newly triggered because we parse the css for opacity *TWICE*.  This is a result of
	incorrectly "invalidating" the style attribute (from fix for 3790449).  The second parse was from
	the cssText() of the style.

        Reviewed by Hyatt.

        * khtml/css/css_valueimpl.cpp:
        (DOM::CSSPrimitiveValueImpl::cssText):


Canonical link: https://commits.webkit.org/7602@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@8658 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/khtml/css/css_valueimpl.cpp
@@ -824,7 +824,7 @@ DOM::DOMString CSSPrimitiveValueImpl::cssText() const
 	    // ###
 	    break;
 	case CSSPrimitiveValue::CSS_NUMBER:
-	    text = DOMString(QString::number( (int)m_value.num ));
+	    text = DOMString(QString::number( m_value.num ));
 	    break;
 	case CSSPrimitiveValue::CSS_PERCENTAGE:
 	    text = DOMString(QString::number( m_value.num ) + "%");
